pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "witness"
version = "0.0.1"
description = "Local-first observability for browser agents. Datadog for browser agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Eric Catalano" }]
keywords = ["browser-use", "agents", "observability", "llm", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "sqlmodel>=0.0.22",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer>=0.12.5",
    "rich>=13.9.0",
    "pydantic>=2.9.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "traceloop-sdk>=0.59.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "posthog>=3.5.0",
]

[project.optional-dependencies]
browser-use = ["browser-use==0.11.13"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.scripts]
witness = "witness.cli:app"

[project.urls]
Homepage = "https://github.com/ericcatalano/witness"

[tool.hatch.build.targets.wheel]
packages = ["witness"]

[tool.hatch.build.targets.wheel.force-include]
"witness/_viewer_dist" = "witness/_viewer_dist"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
